# docs/swagger.yaml
basePath: /v1
definitions:
  models.ChartRepo:
    properties:
      cached:
        type: boolean
      home:
        type: string
      icon:
        type: string
      latest_version:
        type: string
      name:
        type: string
      owner_name:
        type: string
      verify_status:
        type: string
      version_count:
        type: integer
    type: object
  models.ChartVersion:
    properties:
      appVersion:
        type: string
      description:
        type: string
      digest:
        type: string
      name:
        type: string
      url:
        type: string
      version:
        type: string
    type: object
  models.Repository:
    properties:
      cached:
        type: boolean
      description:
        type: string
      owner_name:
        type: string
      pull_count:
        type: integer
      repo_name:
        type: string
      star_count:
        type: integer
      tag_count:
        type: integer
      verify_status:
        type: string
    type: object
  models.RepositoryTag:
    properties:
      architecture:
        type: string
      author:
        type: string
      cached:
        type: boolean
      digest:
        type: string
      docker_version:
        type: string
      os:
        type: string
      os_version:
        type: string
      size:
        type: integer
      tag_name:
        type: string
      verify_status:
        type: string
    type: object
  models.SeederNode:
    properties:
      advertise_addr:
        type: string
      bind_addr:
        type: string
      node_name:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
host: seeder.cloudminds.com
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: Server for image/chart repo consistent.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Example API
  version: "1.0"
paths:
  /api/v1/chart:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: PageSize
        in: query
        name: pageSize
        type: integer
      - description: VerifyStatus
        in: query
        name: status
        type: boolean
      - description: Cached
        in: query
        name: cached
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChartRepo'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取Chart仓库列表
  /api/v1/chart/{repo}/{version}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.ChartVersion'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 删除本地指定Chart仓库的指定版本
  /api/v1/chart/{repo}/{version}/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: Download the version to local
        in: body
        name: version
        required: true
        schema:
          $ref: '#/definitions/models.ChartVersion'
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.ChartVersion'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 下载更新指定Chart仓库的指定版本到本地仓库
  /api/v1/chart/{repo}/charts:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: PageSize
        in: query
        name: pageSize
        type: integer
      - description: VerifyStatus
        in: query
        name: status
        type: boolean
      - description: Cached
        in: query
        name: cached
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.ChartVersion'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取指定Chart仓库的版本列表
  /api/v1/cluster:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: PageSize
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SeederNode'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取edge-cloud整体集群信息
  /api/v1/health:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: '{"code":S200,"data":{},"msg":"ok"}'
          schema:
            type: string
      summary: 获取服务健康状态
  /api/v1/repository:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: PageSize
        in: query
        name: pageSize
        type: integer
      - description: VerifyStatus
        in: query
        name: status
        type: boolean
      - description: Cached
        in: query
        name: cached
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Repository'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取镜像仓库列表
  /api/v1/repository/{repo}/{tag}:
    delete:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.RepositoryTag'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 删除本地指定镜像仓库的指定tag
  /api/v1/repository/{repo}/{tag}/download:
    post:
      consumes:
      - application/json
      parameters:
      - description: Download the tag to local
        in: body
        name: tag
        required: true
        schema:
          $ref: '#/definitions/models.RepositoryTag'
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/models.RepositoryTag'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 下载更新指定镜像仓库的指定tag到本地仓库
  /api/v1/repository/{repo}/tags:
    get:
      consumes:
      - application/json
      parameters:
      - description: Page
        in: query
        name: page
        type: integer
      - description: PageSize
        in: query
        name: pageSize
        type: integer
      - description: VerifyStatus
        in: query
        name: status
        type: boolean
      - description: Cached
        in: query
        name: cached
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.RepositoryTag'
        "500":
          description: Internal Error
          schema:
            type: string
      summary: 获取单个镜像仓库的tag列表
swagger: "2.0"